backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
prophet==1.1.5
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.3
supabase==2.3.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.3
scikit-learn==1.4.0
